Bug: fix -r option behavior
The implementation in the run() function was wrong. What we ended up
doing when we were called as "dotcp -r ..." was
    sudo -AE /path/to/dotcp main
While this does call the script itself again with sudo, we don't pass
any options, such as -s, i.e. simulate did not work. Also, passing
"main" in that case makes no sense at all :) In the end, we did a dotcp
run as root w/o any options. The additional "main" arg is ignored, which
is why no error is thrown. Sorry folks!